% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoembed-package.R
\docType{package}
\name{autoembed-package}
\alias{autoembed}
\alias{autoembed-package}
\title{autoembed: Generate Temporal Lead or Lag Embedding Features}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/schafer-research-lab/autoembed}
  \item Report bugs at \url{https://github.com/schafer-research-lab/autoembed/issues}
}

}
\author{
\strong{Maintainer}: Toryn Schafer \email{toryn27@gmail.com} (\href{https://orcid.org/0000-0001-5594-7697}{ORCID})

Authors:
\itemize{
  \item John Paul Acosta
  \item Brian Park
}

}
\keyword{internal}
