% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automation.R
\name{data.frame_lag_lead}
\alias{data.frame_lag_lead}
\title{Generate temporal lead or lag embeddings}
\usage{
data.frame_lag_lead(
  dataframe,
  covariates,
  nlags = 0,
  nleads = 0,
  vlags = c(),
  vleads = c(),
  grouping = NA,
  skips = 1
)
}
\arguments{
\item{dataframe}{a tidy dataframe}

\item{covariates}{a character vector of strings that specify covariates to lag in the dataframe}

\item{nlags}{integer number of times to lag the covariates (default 0)}

\item{nleads}{integer number of times to lead the covariates (default 0)}

\item{vlags}{vector of integers for creating multiple lags, overrides nlags & skips (default c())}

\item{vleads}{vector of integers for creating multiple leads, overrides nleads & skips (default c())}

\item{grouping}{string specifying column name to group data by to ensure that lagging does not go across groups}

\item{skips}{integer number of entries to skip across each separate lag (default 1)}
}
\value{
A named list with two components

\item{dataframe}{A dataframe with the original columns of the input dataframe and additional columns}
\item{new.covariates}{A character vector of the additional columns names}

The additional column names are created with the following string pattern:
}
\description{
Generate temporal lead or lag embeddings
}
\examples{

df <- data.frame(x = 1:20,
                 y = 20:1)

df_lag <- data.frame_lag_lead(df, covariates = "x")
df_lead <- data.frame_lag_lead(df, covariates = "x", nleads = 1)
}
