% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automation.R
\name{data.frame.lag.lead}
\alias{data.frame.lag.lead}
\title{Title}
\usage{
data.frame.lag.lead(
  dataframe,
  covariates,
  nlags = 0,
  nleads = 0,
  vlags = c(),
  vleads = c(),
  grouping = NA,
  skips = 1
)
}
\arguments{
\item{skips}{}
}
\description{
Title
}
